Added validation of api_key.
Before you could enter any character as the api_key, causing directory traversal in the request to PLS. So if you entered ../group? then the request to PLS would return all the permission groups as a list. Since no validation was performed whether the returned json was a list or a dict the check that 'login' was in the returned JSON always succeded. Now the api_key is restricted to a-zA-Z0-9-_=