Added validation of callback url so it's always a sumdomain of datasektionen.se
Without this validation it is possible to leak the token of any logged in user by having them visit a malicious site. This validation is right now performed with a regex but should preferably be just a whitelist, since D-sys insisted on that it would be less maintenance. This validation is turned off when developing locally.