Attempt to upgrade internal callback URL to HTTPS
We never want to risk sending the login token via an insecure connection, so making this an explicit step makes sense. Most (if not all) of our services are configured to use https, and the ones that don't, or do but do so improperly, should not be allowed to use login anyway.